portuense-backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.1.0
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.9.0
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.1
